routes added for manage categories and manage roles
