+++ apps/goals-on-track/pages/dashboard/manage/categories/index.module.scss
@@ -1,0 +1,10 @@
+.container{
+    .dashboard_page_container{
+        display: flex;
+
+        .header_and_user_detail_section{
+            width: 100%;
+        }
+    }
+}
+
