+++ libs/shared/data-access/src/lib/api/allRolesApi.json
@@ -24,19 +24,9 @@
       "id": "JaSwJis"
     },
     {
-      "name": "x",
-      "description": "xxx",
-      "id": "Nc0myRv"
-    },
-    {
       "name": "xxx",
       "description": "xxx",
       "id": "K-0dfKS"
-    },
-    {
-      "name": "vishal",
-      "description": "vishal",
-      "id": "kqCl_d5"
     }
   ]
 }
